Doctrine ORM Manager, Admin Class, Unit Tests
Switched from manually doing all the SQL to abstracting it with an ORM
called Doctrine 2.  Doctrine generates SQL using some commandline
utilities, so you have to do a little work to keep the database
updated.  I’ll write a makefile at some point to automate that.
There’s also the Admin and AdminManager (which handles the storage and
retrieval of Admin objects) classes, and a set of unit tests.