Integrated the certificate generation
Adam’s certificate generation was great, but it wasn’t hooked into the
backend database and file services.  This uses the same certificate
generation code, but wires it into the existing pages.

There’s more to do in terms of improving the UI, but the TeX jobs gets
fired off as an async process now.  The user *can* encounter a race
condition where the file hasn’t been generated yet, but that’s a
problem for another day.