Bugfix: Award By Granter showing odd results
I changed the method signature for
AwardManager->getAwardCountByGranter() to facilitate pagination on the
create award pages where we’re limiting results to a specific granter,
but I forgot to update this call, so it was pushing a date object into
the granterId variable when the dates weren’t null.  The the dates were
null, everything defaulted correctly, so the results looked fine.